refactor: 관심사에 따라 훅 분리하기
